refactor: move mpv-related composbles to `components/MultiPageViewer/`
