+++ src/components/MultiPageViewer/composables/ImageMagnifier/useMagnifierGesture.ts
