+++ src/components/MultiPageViewer/composables/ImageMagnifier/useMagnifierImageLoader.ts
