+++ src/components/MultiPageViewer/composables/ImageMagnifier/useMagnifierStyle.ts
